Create R_script
This script contains the steps to model substrate-dependent active antagonism in microbial co-cultures using a mass-balanced framework.  In this script, we employ this framework on data from Vazquez-Munoz et al. (2022), which describes a liquid co-culture of a Lactobacillus species and Candida albicans. This co-culture's dynamics resulted from resource competition, cross-feeding as well as active antagonism via the secretion of a soluble anticandidal metabolite. These multimodal interactions were explored in this model script by means of a system of coupled ordinary differential equations.